refactored auth file structure
