+++ src/pages/Register/Register.css
@@ -1,0 +1,37 @@
+.auth-container {
+    display: flex;
+    justify-content: center;
+    align-content: center;
+    flex-flow: column wrap;
+}
+
+.auth-container form {
+    display: flex;
+    flex-flow: column nowrap;
+    justify-content: space-evenly;
+    align-content: space-evenly;
+    border-style: solid;
+    border-width: 0.1rem;
+    border-color: white;
+    border-radius: 0.5rem;
+    padding: 2rem;
+    gap: 0.5rem;
+}
+
+.auth-container h2 {
+    color: white;
+    text-align: center;
+}
+
+.auth-container label {
+    margin-top: 1rem;
+    font-weight: bold;
+}
+
+.auth-container input {
+    margin-top: 0.5rem;
+}
+
+.auth-container .error {
+    color: red;
+}
